Big Changes
Added data storage system. The mod will now store data about nodes in the Assets slot of the world. It is all non-persistent. The main goal of this is to optimize memory usage by allowing Neos to take care of the memory management of this data so I don't have to do it myself. It also allows for a much better auto-refresh feature for when you change the config. The nodes will no longer regenerate their whole visual, but will only update the color using stored refs to IFields. The auto-refresh feature is now enabled by default because it isn't super slow anymore.